# Conflicts:
#	Gems/Terrain/Code/Source/TerrainRenderer/TerrainFeatureProcessor.cpp

+++ Gems/Multiplayer/Code/Source/Editor/MultiplayerEditorSystemComponent.cpp
@@ -19,9 +19,11 @@
 #include <AzCore/Interface/Interface.h>
 #include <AzCore/Serialization/SerializeContext.h>
 #include <AzCore/Utils/Utils.h>
+#include <AzFramework/API/ApplicationAPI.h>
 #include <AzNetworking/Framework/INetworking.h>
 #include <AzToolsFramework/Entity/PrefabEditorEntityOwnershipInterface.h>
 #include <Atom/RPI.Public/RPISystemInterface.h>
+#include <AzCore/Settings/SettingsRegistryMergeUtils.h>
 
 namespace Multiplayer
 {
@@ -193,22 +195,62 @@ namespace Multiplayer
         }
     }
 
+    bool MultiplayerEditorSystemComponent::FindServerLauncher(AZ::IO::FixedMaxPath& serverPath)
+    {
+        serverPath.clear();
+
+        // 1. Try the path from `editorsv_process` cvar.
+        const AZ::IO::FixedMaxPath serverPathFromCvar{ AZ::CVarFixedString(editorsv_process).c_str()};
+        if (AZ::IO::SystemFile::Exists(serverPathFromCvar.c_str()))
+        {
+            serverPath = serverPathFromCvar;
+            return true;
+        }
+
+        // 2. Try from the executable folder where the Editor was launched from.
+        AZ::IO::FixedMaxPath serverPathFromEditorLocation = AZ::Utils::GetExecutableDirectory();
+        serverPathFromEditorLocation /= AZStd::string_view(AZ::Utils::GetProjectName() + ".ServerLauncher" + AZ_TRAIT_OS_EXECUTABLE_EXTENSION);
+        if (AZ::IO::SystemFile::Exists(serverPathFromEditorLocation.c_str()))
+        {
+            serverPath = serverPathFromEditorLocation;
+            return true;
+        }
+
+        // 3. Try from the project's build folder.
+        AZ::IO::FixedMaxPath serverPathFromProjectBin;
+        if (const auto settingsRegistry = AZ::SettingsRegistry::Get(); settingsRegistry != nullptr)
+        {
+            if (AZ::IO::FixedMaxPath projectModulePath;
+                settingsRegistry->Get(projectModulePath.Native(), AZ::SettingsRegistryMergeUtils::FilePathKey_ProjectConfigurationBinPath))
+            {
+                serverPathFromProjectBin /= projectModulePath;
+                serverPathFromProjectBin /= AZStd::string_view(AZ::Utils::GetProjectName() + ".ServerLauncher" + AZ_TRAIT_OS_EXECUTABLE_EXTENSION);
+                if (AZ::IO::SystemFile::Exists(serverPathFromProjectBin.c_str()))
+                {
+                    serverPath = serverPathFromProjectBin;
+                    return true;
+                }
+            }
+        }
+
+        AZ_Error(
+            "MultiplayerEditor", false,
+            "The ServerLauncher binary is missing! (%s), (%s) and (%s) were tried. Please build server launcher or specify using editorsv_process.",
+            serverPathFromCvar.c_str(),
+            serverPathFromEditorLocation.c_str(),
+            serverPathFromProjectBin.c_str());
+
+        return false;
+    }
+
     void MultiplayerEditorSystemComponent::LaunchEditorServer()
     {
         // Assemble the server's path
-        AZ::CVarFixedString serverProcess = editorsv_process;
         AZ::IO::FixedMaxPath serverPath;
-        if (serverProcess.empty())
+        if (!FindServerLauncher(serverPath))
         {
-            // If enabled but no process name is supplied, try this project's ServerLauncher
-            serverProcess = AZ::Utils::GetProjectName() + ".ServerLauncher";
-            serverPath = AZ::Utils::GetExecutableDirectory();
-            serverPath /= serverProcess + AZ_TRAIT_OS_EXECUTABLE_EXTENSION;
-        }
-        else
-        {
-            serverPath = serverProcess;
-        }
+            return;
+        }        
 
         // Start the configured server if it's available
         AzFramework::ProcessLauncher::ProcessLaunchInfo processLaunchInfo;
@@ -220,22 +262,12 @@ namespace Multiplayer
             server_rhi = static_cast<AZ::CVarFixedString>(editorsv_rhi_override);
         }
 
-        const auto console = AZ::Interface<AZ::IConsole>::Get();
-        AZ::CVarFixedString sv_defaultPlayerSpawnAsset;
-
-        if (console->GetCvarValue("sv_defaultPlayerSpawnAsset", sv_defaultPlayerSpawnAsset) != AZ::GetValueResult::Success)
-        {
-            AZ_Assert( false,
-                "MultiplayerEditorSystemComponent::LaunchEditorServer failed! Could not find the sv_defaultPlayerSpawnAsset cvar; the editor-server "
-                "will fall back to using some other default player! Please update this code to use a valid cvar!")
-        }
-        
         processLaunchInfo.m_commandlineParameters = AZStd::string::format(
-            R"("%s" --project-path "%s" --editorsv_isDedicated true --sv_defaultPlayerSpawnAsset "%s" --rhi "%s")",
+            R"("%s" --project-path "%s" --editorsv_isDedicated true --bg_ConnectToAssetProcessor false --rhi "%s" --editorsv_port %i)",
             serverPath.c_str(),
             AZ::Utils::GetProjectPath().c_str(),
-            sv_defaultPlayerSpawnAsset.c_str(),
-            server_rhi.GetCStr()
+            server_rhi.GetCStr(),
+            static_cast<uint16_t>(editorsv_port)
         );
         processLaunchInfo.m_showWindow = true;
         processLaunchInfo.m_processPriority = AzFramework::ProcessPriority::PROCESSPRIORITY_NORMAL;
@@ -244,19 +276,22 @@ namespace Multiplayer
         AzFramework::ProcessWatcher* outProcess = AzFramework::ProcessWatcher::LaunchProcess(
             processLaunchInfo, AzFramework::ProcessCommunicationType::COMMUNICATOR_TYPE_STDINOUT);
 
-        AZ_Error(
-            "MultiplayerEditor", processLaunchInfo.m_launchResult != AzFramework::ProcessLauncher::ProcessLaunchResult::PLR_MissingFile,
-            "LaunchEditorServer failed! The ServerLauncher binary is missing! (%s)  Please build server launcher.", serverPath.c_str())
-
-        // Stop the previous server if one exists
-        if (m_serverProcessWatcher)
+        if (outProcess)
         {
-            AZ::TickBus::Handler::BusDisconnect();
-            m_serverProcessWatcher->TerminateProcess(0);
+            // Stop the previous server if one exists
+            if (m_serverProcessWatcher)
+            {
+                AZ::TickBus::Handler::BusDisconnect();
+                m_serverProcessWatcher->TerminateProcess(0);
+            }
+            m_serverProcessWatcher.reset(outProcess);
+            m_serverProcessTracePrinter = AZStd::make_unique<ProcessCommunicatorTracePrinter>(m_serverProcessWatcher->GetCommunicator(), "EditorServer");
+            AZ::TickBus::Handler::BusConnect();
         }
-        m_serverProcessWatcher.reset(outProcess);
-        m_serverProcessTracePrinter = AZStd::make_unique<ProcessCommunicatorTracePrinter>(m_serverProcessWatcher->GetCommunicator(), "EditorServer");
-        AZ::TickBus::Handler::BusConnect();
+        else
+        {
+            AZ_Error("MultiplayerEditor", outProcess, "LaunchEditorServer failed! Unable to create AzFramework::ProcessWatcher.");
+        }
     }
 
     void MultiplayerEditorSystemComponent::OnGameEntitiesStarted()
@@ -272,7 +307,9 @@ namespace Multiplayer
         auto prefabEditorEntityOwnershipInterface = AZ::Interface<AzToolsFramework::PrefabEditorEntityOwnershipInterface>::Get();
         if (!prefabEditorEntityOwnershipInterface)
         {
-            AZ_Error("MultiplayerEditor", prefabEditorEntityOwnershipInterface != nullptr, "PrefabEditorEntityOwnershipInterface unavailable");
+            bool prefabSystemEnabled = false;
+            AzFramework::ApplicationRequests::Bus::BroadcastResult(prefabSystemEnabled, &AzFramework::ApplicationRequests::IsPrefabSystemEnabled);
+            AZ_Error("MultiplayerEditor", !prefabSystemEnabled, "PrefabEditorEntityOwnershipInterface unavailable but prefabs are enabled");
             return;
         }
 
@@ -361,7 +398,9 @@ namespace Multiplayer
         const auto prefabEditorEntityOwnershipInterface = AZ::Interface<AzToolsFramework::PrefabEditorEntityOwnershipInterface>::Get();
         if (!prefabEditorEntityOwnershipInterface)
         {
-            AZ_Error("MultiplayerEditor", prefabEditorEntityOwnershipInterface != nullptr, "PrefabEditorEntityOwnershipInterface unavailable")
+            bool prefabSystemEnabled = false;
+            AzFramework::ApplicationRequests::Bus::BroadcastResult(prefabSystemEnabled, &AzFramework::ApplicationRequests::IsPrefabSystemEnabled);
+            AZ_Error("MultiplayerEditor", !prefabSystemEnabled, "PrefabEditorEntityOwnershipInterface unavailable but prefabs are enabled");
             return;
         }
